extract locking queue writing logic
